+++ packages/examples/src/mediaQueries/index.js
@@ -1,5 +1,10 @@
 import React from 'react';
-import { Document, Page, View, StyleSheet } from '@react-pdf/renderer-react17support';
+import {
+  Document,
+  Page,
+  View,
+  StyleSheet,
+} from '@andfaulkner/react-pdf-renderer-react17support';
 
 const styles = StyleSheet.create({
   body: {