+++ packages/examples/src/resume/Header.js
@@ -1,6 +1,11 @@
 import React from 'react';
 
-import { Link, Text, View, StyleSheet } from '@react-pdf/renderer-react17support';
+import {
+  Link,
+  Text,
+  View,
+  StyleSheet,
+} from '@andfaulkner/react-pdf-renderer-react17support';
 
 const styles = StyleSheet.create({
   container: {