+++ packages/examples/src/svg/Svg1.js
@@ -1,5 +1,12 @@
 import React from 'react';
-import { Rect, Svg, G, Path, Text, Tspan } from '@react-pdf/renderer-react17support';
+import {
+  Rect,
+  Svg,
+  G,
+  Path,
+  Text,
+  Tspan,
+} from '@andfaulkner/react-pdf-renderer-react17support';
 
 const Svg1 = () => (
   <Svg width="400" viewBox="0 0 600 400">