+++ packages/examples/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @react-pdf/examples-react17support
+# @andfaulkner/react-pdf-examples-react17support
 
 ## 3.0.0
 
@@ -7,7 +7,7 @@
 - Add React 17 support (removes React 16 support).
 
 - Updated dependencies
-  - @react-pdf/renderer-react17support@4.0.0
+  - @andfaulkner/react-pdf-renderer-react17support@4.0.0
 
 ## 2.0.2